requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
spotipy==2.23.0
python-multipart==0.0.9
langdetect==1.0.9
google-search-results==2.4.2
jinja2==3.1.3
requests==2.31.0 